Add PartyModel
We have to create our own PartyModel as for fuy's purposes we'd like to know to the UUID of party members. Wynntils' PartyModel is also seemingly out of date? I'm unsure whether this covers all cases (as testing it requires me DM a friend! and they are usually not keen on rapid iteration).